test(apple): share one wait helper instead of sleeping
