+++ swift/apple/FirezoneKit/Tests/FirezoneKitTests/CancellableTaskTests.swift
@@ -10,24 +10,8 @@ private actor Flag {
     value = newValue
   }
 
-  /// The flag once it is set, or its value at `timeout`.
-  ///
-  /// What these tests wait on is a task noticing cancellation, which takes as
-  /// long as the scheduler needs. A fixed sleep either gives that away or, on a
-  /// loaded machine, ends first and reports a failure that says nothing about
-  /// the code under test.
-  func waitUntilSet(timeout: Duration = .seconds(5)) async -> Bool {
-    let deadline = ContinuousClock.now.advanced(by: timeout)
-
-    while ContinuousClock.now < deadline {
-      if value {
-        return true
-      }
-
-      try? await Task.sleep(for: .milliseconds(5))
-    }
-
-    return value
+  func get() -> Bool {
+    value
   }
 }
 
@@ -35,22 +19,21 @@ private actor Flag {
 struct CancellableTaskTests {
 
   @Test("Task executes its operation")
-  func taskExecutesOperation() async {
+  func taskExecutesOperation() async throws {
     let executed = Flag()
 
     let task = CancellableTask {
       await executed.set(true)
     }
 
-    let wasExecuted = await executed.waitUntilSet()
-    #expect(wasExecuted == true)
+    try await waitUntil { await executed.get() }
 
     // Keep task alive until assertion
     _ = task
   }
 
   @Test("Task is cancelled when CancellableTask is deallocated")
-  func taskCancelledOnDealloc() async {
+  func taskCancelledOnDealloc() async throws {
     let wasCancelled = Flag()
 
     do {
@@ -64,12 +47,11 @@ struct CancellableTaskTests {
       // CancellableTask goes out of scope here, triggering deinit -> cancel
     }
 
-    let taskWasCancelled = await wasCancelled.waitUntilSet()
-    #expect(taskWasCancelled == true)
+    try await waitUntil { await wasCancelled.get() }
   }
 
   @Test("Setting to nil cancels the task")
-  func settingToNilCancelsTask() async {
+  func settingToNilCancelsTask() async throws {
     let wasCancelled = Flag()
 
     var task: CancellableTask? = CancellableTask {
@@ -83,8 +65,7 @@ struct CancellableTaskTests {
     // Set to nil, triggering cancellation
     task = nil
 
-    let taskWasCancelled = await wasCancelled.waitUntilSet()
-    #expect(taskWasCancelled == true)
+    try await waitUntil { await wasCancelled.get() }
 
     // Silence unused variable warning
     _ = task
